+++ Modules/Config/ConfigModule.lua
@@ -388,7 +388,6 @@ do
                  "showHoveredPOI",
                  "lootFilterUpgrades",
                  "colorWarbandBonus",
-                 "disableTaintWorkarounds",
                  "enableDebugging"
             }